Run the workflow tutorial and its tests where their optional pieces are not
Three things the branch had not been through CI to find. The tutorial's
save and load example defines its task classes in the page, which the
generated doc tests run inside a function, where a class is a local one and
no document could name it -- so that chunk is not mirrored into them; quarto
still runs it. The tests which save YAML skip where pyyaml is not installed,
as the free-threaded build is one place it is not. And the pinned
fingerprint of an array parameter spells its dtype out, since `arange` gives
int32 where a C long is 32 bits and the digest names the dtype.

+++ tests/test_workflow/test_task.py
@@ -132,7 +132,11 @@ class TestFingerprint:
         cannot see a change in how a tagged one -- an array, a time -- is
         written; this pins that half.
         """
-        task_ = TimedValueTask(when=np.datetime64("2020-01-01"), where=np.arange(3))
+        # The dtype is spelled out: `arange` gives int32 where a C long is
+        # 32 bits -- WebAssembly, for one -- and the digest names the dtype.
+        task_ = TimedValueTask(
+            when=np.datetime64("2020-01-01"), where=np.arange(3, dtype="int64")
+        )
         assert task_.fingerprint == "53791ea82cf3fe6d"
 
     def test_params_matter(self):
